1. remove userinfo 2. hide sort in some fields 3. fix operate
