+++ src/i18n/en.json
@@ -21,6 +21,7 @@
   "search": "Search",
   "name": "name",
   "roles": "roles",
+  "operate": "operate",
   "base": {
     "confirm": {
       "delete": "Are you sure?",
@@ -81,19 +82,6 @@
       }
     },
     "User": {
-      "attr": {
-        "id": "id",
-        "username": "user name",
-        "cellphone": "cell phone",
-        "status": "status",
-        "role": "role",
-        "desc": "desc",
-        "createdAt": "registered at",
-        "updatedAt": "updated at"
-      },
-      "management": "user management"
-    },
-    "Userinfo": {
       "attr": {
         "id": "id",
         "username": "user name",
