+++ src/views/lesson/teacher_cdkey/crud.vue
@@ -4,7 +4,7 @@
       <el-button class="filter-item" style="margin-left: 10px;" @click="handleGenerate" type="primary">Generate Code</el-button>
     </div>
 
-    <crud-table :key="nestedKey" :listLoading="listLoading" :resourceClass="resourceClass" :list="list" :filter="colFilter" @handleAction="handleAction" @handleSort="handleSort"> </crud-table>
+    <crud-table :listLoading="listLoading" :resourceClass="resourceClass" :list="list" :filter="colFilter" @handleAction="handleAction" @handleSort="handleSort"></crud-table>
 
     <crud-paginate :listQuery="listQuery" :total="total" @handleSizeChange="handleSizeChange" @handleCurrentChange="handleCurrentChange"> </crud-paginate>
 
@@ -15,7 +15,6 @@
         <el-button type="primary" @click="generateData">OK</el-button>
       </div>
     </el-dialog>
-
   </div>
 </template>
